# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: KnowBound-CJ API
  version: 0.2.0
  description: Contract-first API for the KnowBound-CJ learning assistant.
servers:
- url: http://localhost:8000
paths:
...
  /courses:
    get:
      operationId: listCourses
      summary: List courses visible to the authenticated actor.
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Enrolled courses for a student, or owned/authorized courses for a teacher.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseSummary'
        '401':
          description: Bearer token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /courses/{course_id}/exercises:
    get:
      operationId: listCourseExercises
      summary: List exercises after server-side course membership checks.
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/CourseId'
      responses:
        '200':
          description: Exercises in the authorized course.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExerciseSummary'
        '401':
          description: Bearer token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Returns RESOURCE_NOT_AVAILABLE for both unknown and unauthorized course IDs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /teacher/courses:
    post:
      operationId: createTeacherCourse
      summary: Create a course owned by the authenticated teacher.
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseRequest'
      responses:
        '201':
          description: Course created with ownership derived from the bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseSummary'
        '401':
          description: Bearer token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Student actors receive RESOURCE_NOT_AVAILABLE.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/InvalidRequest'
  /courses/{course_id}/enrollments:
    post:
      operationId: enrollCurrentStudent
      summary: Enroll the authenticated student; no user ID is accepted from the client.
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/CourseId'
      responses:
        '201':
          description: Current student enrolled in the course.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enrollment'
        '401':
          description: Bearer token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Returns RESOURCE_NOT_AVAILABLE for an unknown course or a disallowed actor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Role:
      type: string
      enum:
      - student
      - teacher
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
      - code
      - message
      - request_id
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
...
    CourseSummary:
      type: object
      additionalProperties: false
      required:
      - course_id
      - title
      properties:
        course_id:
          type: string
        title:
          type: string
    ExerciseSummary:
      type: object
      additionalProperties: false
      required:
      - exercise_id
      - course_id
      - title
      properties:
        exercise_id:
          type: string
        course_id:
          type: string
        title:
          type: string
    CreateCourseRequest:
      type: object
      additionalProperties: false
      required:
      - title
      properties:
        title:
          type: string
    Enrollment:
      type: object
      additionalProperties: false
      required:
      - course_id
      - user_id
      - role
      properties:
        course_id:
          type: string
        user_id:
          type: string
        role:
          $ref: '#/components/schemas/Role'
...
  parameters:
    CourseId:
      name: course_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    InvalidRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'